pygento/requirements.txt
google-genai>=0.1.0
requests>=2.31.0
Pillow>=10.0.0
flask>=2.0.0
